Updated model decimal fields

+++ app/models.py
@@ -17,7 +17,8 @@ class Watchlist(models.Model):
 
 class Cash(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
-    cash_balance = models.IntegerField(default=1000000)
+    cash_balance = models.DecimalField(max_digits=19, decimal_places=2, default=1000000.00)
+    # cash_balance = models.IntegerField(default=1000000)
 
     def __str__(self):
         return f"{self.user} has {self.cash_balance} remaining"
@@ -26,8 +27,10 @@ class Holdings(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     symbol = models.CharField(max_length=10)
     quantity = models.IntegerField()
-    avg_price = models.IntegerField()
-    total_amount = models.IntegerField()
+    avg_price = models.DecimalField(max_digits=19, decimal_places=2)
+    # avg_price = models.IntegerField()
+    # total_amount = models.IntegerField()
+    total_amount = models.DecimalField(max_digits=19, decimal_places=2)
 
     def __str__(self):
         return f"{self.user} has {self.quantity} shares of {self.symbol}"
@@ -36,8 +39,10 @@ class Papertrading(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     symbol = models.CharField(max_length=10)
     quantity = models.IntegerField()
-    price = models.IntegerField()
-    total_amount = models.IntegerField()
+    price = models.DecimalField(max_digits=19, decimal_places=2)
+    # price = models.IntegerField()
+    total_amount = models.DecimalField(max_digits=19, decimal_places=2)
+    # total_amount = models.IntegerField()
     direction = models.CharField(max_length=4)
 
     def __str__(self):
